agent/pyproject.toml
[tool.poetry]
name = "mcp-agent"
version = "0.1.0"
description = "MCP Agent implementation"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
python-dotenv = "^1.0.1"
langchain = "^0.1.0"
langchain-core = "^0.3.36"
langchain-google-genai = "^0.0.11"
pydantic = "^2.7.2"
websockets = "^12.0"
langgraph = "^0.3.7"
google-generativeai = "^0.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 